+++ Epoch/include/Epoch/Graphics/Renderer.h
@@ -11,6 +11,7 @@ namespace Epoch {
 	{
 	public:
 		static void init();
+		static void onWindowResized(uint32_t width, uint32_t height);
 
 		static void beginScene(OrthographicCamera& camera);
 		static void endScene();