+++ Epoch/include/Epoch/Platform/OpenGL/OpenGLRendererAPI.h
@@ -8,6 +8,7 @@ namespace Epoch {
 	{
 	public:
 		virtual void init() override;
+		virtual void setViewPort(uint32_t x, uint32_t y, uint32_t width, uint32_t height) override;
 
 		virtual void setClearColor(const glm::vec4& color) override;
 		virtual void clear() override;