+++ Epoch/include/Epoch/Platform/OpenGL/OpenGLShader.h
@@ -10,7 +10,7 @@ namespace Epoch {
 	{
 	public:
 		OpenGLShader(const std::string& filepath);
-		OpenGLShader(const std::string& vertexSrc, const std::string& fragmentSrc);
+		OpenGLShader(const std::string& name, const std::string& vertexSrc, const std::string& fragmentSrc);
 		~OpenGLShader();
 
 		virtual void bind() const override;