+++ docs/Changelog.md
@@ -31,3 +31,7 @@ Note: I am a college student this is a passion project which i do in my spare ti
 - Textures
 - Logo
 - CameraController Class
+
+### 9/29/19 ###
+- Shader Library
+- Fullscreen